pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.14",
    "google>=3.0.0",
    "googleapis-common-protos>=1.70.0",
    "grpcio>=1.73.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.69",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "protobuf>=6.31.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
